requirements.txt
discord.py>=2.0.0
yt-dlp>=2023.0.0
PyNaCl>=1.4.0
python-dotenv>=0.19.0 